+++ src/any.rs
@@ -1,10 +1,8 @@
 //! Helper module which defines the `Any` trait to to allow dynamic value handling.
 
+use crate::module::Module;
 use crate::parser::{ImmutableString, INT};
 use crate::r#unsafe::{unsafe_cast_box, unsafe_try_cast};
-
-#[cfg(not(feature = "no_module"))]
-use crate::module::Module;
 
 #[cfg(not(feature = "no_float"))]
 use crate::parser::FLOAT;
@@ -160,7 +158,6 @@ pub enum Union {
     Array(Box<Array>),
     #[cfg(not(feature = "no_object"))]
     Map(Box<Map>),
-    #[cfg(not(feature = "no_module"))]
     Module(Box<Module>),
     Variant(Box<Box<dyn Variant>>),
 }
@@ -198,7 +195,6 @@ impl Dynamic {
             Union::Array(_) => TypeId::of::<Array>(),
             #[cfg(not(feature = "no_object"))]
             Union::Map(_) => TypeId::of::<Map>(),
-            #[cfg(not(feature = "no_module"))]
             Union::Module(_) => TypeId::of::<Module>(),
             Union::Variant(value) => (***value).type_id(),
         }
@@ -218,7 +214,6 @@ impl Dynamic {
             Union::Array(_) => "array",
             #[cfg(not(feature = "no_object"))]
             Union::Map(_) => "map",
-            #[cfg(not(feature = "no_module"))]
             Union::Module(_) => "sub-scope",
 
             #[cfg(not(feature = "no_std"))]
@@ -242,7 +237,6 @@ impl fmt::Display for Dynamic {
             Union::Array(value) => write!(f, "{:?}", value),
             #[cfg(not(feature = "no_object"))]
             Union::Map(value) => write!(f, "#{:?}", value),
-            #[cfg(not(feature = "no_module"))]
             Union::Module(value) => write!(f, "{:?}", value),
 
             #[cfg(not(feature = "no_std"))]
@@ -266,7 +260,6 @@ impl fmt::Debug for Dynamic {
             Union::Array(value) => write!(f, "{:?}", value),
             #[cfg(not(feature = "no_object"))]
             Union::Map(value) => write!(f, "#{:?}", value),
-            #[cfg(not(feature = "no_module"))]
             Union::Module(value) => write!(f, "{:?}", value),
 
             #[cfg(not(feature = "no_std"))]
@@ -290,7 +283,6 @@ impl Clone for Dynamic {
             Union::Array(ref value) => Self(Union::Array(value.clone())),
             #[cfg(not(feature = "no_object"))]
             Union::Map(ref value) => Self(Union::Map(value.clone())),
-            #[cfg(not(feature = "no_module"))]
             Union::Module(ref value) => Self(Union::Module(value.clone())),
             Union::Variant(ref value) => (***value).clone_into_dynamic(),
         }
@@ -426,7 +418,6 @@ impl Dynamic {
             Union::Array(value) => unsafe_cast_box::<_, T>(value).ok().map(|v| *v),
             #[cfg(not(feature = "no_object"))]
             Union::Map(value) => unsafe_cast_box::<_, T>(value).ok().map(|v| *v),
-            #[cfg(not(feature = "no_module"))]
             Union::Module(value) => unsafe_cast_box::<_, T>(value).ok().map(|v| *v),
             Union::Variant(value) => (*value).as_box_any().downcast().map(|x| *x).ok(),
         }
@@ -470,7 +461,6 @@ impl Dynamic {
             Union::Array(value) => *unsafe_cast_box::<_, T>(value).unwrap(),
             #[cfg(not(feature = "no_object"))]
             Union::Map(value) => *unsafe_cast_box::<_, T>(value).unwrap(),
-            #[cfg(not(feature = "no_module"))]
             Union::Module(value) => *unsafe_cast_box::<_, T>(value).unwrap(),
             Union::Variant(value) => (*value).as_box_any().downcast().map(|x| *x).unwrap(),
         }
@@ -498,7 +488,6 @@ impl Dynamic {
             Union::Array(value) => (value.as_ref() as &dyn Any).downcast_ref::<T>(),
             #[cfg(not(feature = "no_object"))]
             Union::Map(value) => (value.as_ref() as &dyn Any).downcast_ref::<T>(),
-            #[cfg(not(feature = "no_module"))]
             Union::Module(value) => (value.as_ref() as &dyn Any).downcast_ref::<T>(),
             Union::Variant(value) => value.as_ref().as_ref().as_any().downcast_ref::<T>(),
         }
@@ -524,7 +513,6 @@ impl Dynamic {
             Union::Array(value) => (value.as_mut() as &mut dyn Any).downcast_mut::<T>(),
             #[cfg(not(feature = "no_object"))]
             Union::Map(value) => (value.as_mut() as &mut dyn Any).downcast_mut::<T>(),
-            #[cfg(not(feature = "no_module"))]
             Union::Module(value) => (value.as_mut() as &mut dyn Any).downcast_mut::<T>(),
             Union::Variant(value) => value.as_mut().as_mut_any().downcast_mut::<T>(),
         }