+++ src/engine.rs
@@ -4,7 +4,7 @@ use crate::any::{Dynamic, Union};
 use crate::calc_fn_hash;
 use crate::error::ParseErrorType;
 use crate::fn_native::{CallableFunction, FnCallArgs, Shared};
-use crate::module::Module;
+use crate::module::{resolvers, Module, ModuleResolver};
 use crate::optimize::OptimizationLevel;
 use crate::packages::{CorePackage, Package, PackageLibrary, PackagesCollection, StandardPackage};
 use crate::parser::{Expr, FnAccess, FnDef, ImmutableString, ReturnType, Stmt, AST, INT};
@@ -12,13 +12,10 @@ use crate::r#unsafe::{unsafe_cast_var_name_to_lifetime, unsafe_mut_cast_to_lifet
 use crate::result::EvalAltResult;
 use crate::scope::{EntryType as ScopeEntryType, Scope};
 use crate::token::Position;
-use crate::utils::StaticVec;
+use crate::utils::{StaticVec, StraightHasherBuilder};
 
 #[cfg(not(feature = "no_float"))]
 use crate::parser::FLOAT;
-
-#[cfg(not(feature = "no_module"))]
-use crate::module::{resolvers, ModuleResolver};
 
 use crate::stdlib::{
     any::TypeId,
@@ -196,7 +193,7 @@ impl State {
 ///
 /// The key of the `HashMap` is a `u64` hash calculated by the function `calc_fn_hash`.
 #[derive(Debug, Clone, Default)]
-pub struct FunctionsLib(HashMap<u64, Shared<FnDef>>);
+pub struct FunctionsLib(HashMap<u64, Shared<FnDef>, StraightHasherBuilder>);
 
 impl FunctionsLib {
     /// Create a new `FunctionsLib` from a collection of `FnDef`.
@@ -261,7 +258,7 @@ impl From<Vec<(u64, Shared<FnDef>)>> for FunctionsLib {
 }
 
 impl Deref for FunctionsLib {
-    type Target = HashMap<u64, Shared<FnDef>>;
+    type Target = HashMap<u64, Shared<FnDef>, StraightHasherBuilder>;
 
     fn deref(&self) -> &Self::Target {
         &self.0
@@ -269,7 +266,7 @@ impl Deref for FunctionsLib {
 }
 
 impl DerefMut for FunctionsLib {
-    fn deref_mut(&mut self) -> &mut HashMap<u64, Shared<FnDef>> {
+    fn deref_mut(&mut self) -> &mut HashMap<u64, Shared<FnDef>, StraightHasherBuilder> {
         &mut self.0
     }
 }
@@ -297,7 +294,6 @@ pub struct Engine {
     pub(crate) packages: PackagesCollection,
 
     /// A module resolution service.
-    #[cfg(not(feature = "no_module"))]
     pub(crate) module_resolver: Option<Box<dyn ModuleResolver>>,
 
     /// A hashmap mapping type names to pretty-print names.
@@ -350,8 +346,7 @@ impl Default for Engine {
             #[cfg(not(feature = "no_module"))]
             #[cfg(not(feature = "no_std"))]
             module_resolver: Some(Box::new(resolvers::FileModuleResolver::new())),
-            #[cfg(not(feature = "no_module"))]
-            #[cfg(feature = "no_std")]
+            #[cfg(any(feature = "no_module", feature = "no_std"))]
             module_resolver: None,
 
             type_names: HashMap::new(),
@@ -441,34 +436,32 @@ fn search_scope<'s, 'a>(
         Expr::Variable(x) => x.as_ref(),
         _ => unreachable!(),
     };
-    let index = if state.always_search { None } else { *index };
 
-    #[cfg(not(feature = "no_module"))]
-    {
-        if let Some(modules) = modules.as_ref() {
-            let module = if let Some(index) = modules.index() {
-                scope
-                    .get_mut(scope.len() - index.get())
-                    .0
-                    .downcast_mut::<Module>()
-                    .unwrap()
-            } else {
-                let (id, root_pos) = modules.get(0);
+    if let Some(modules) = modules.as_ref() {
+        let module = if let Some(index) = modules.index() {
+            scope
+                .get_mut(scope.len() - index.get())
+                .0
+                .downcast_mut::<Module>()
+                .unwrap()
+        } else {
+            let (id, root_pos) = modules.get(0);
 
-                scope.find_module(id).ok_or_else(|| {
-                    Box::new(EvalAltResult::ErrorModuleNotFound(id.into(), *root_pos))
-                })?
-            };
+            scope
+                .find_module_internal(id)
+                .ok_or_else(|| Box::new(EvalAltResult::ErrorModuleNotFound(id.into(), *root_pos)))?
+        };
 
-            return Ok((
-                module.get_qualified_var_mut(name, *hash_var, *pos)?,
-                name,
-                // Module variables are constant
-                ScopeEntryType::Constant,
-                *pos,
-            ));
-        }
+        return Ok((
+            module.get_qualified_var_mut(name, *hash_var, *pos)?,
+            name,
+            // Module variables are constant
+            ScopeEntryType::Constant,
+            *pos,
+        ));
     }
+
+    let index = if state.always_search { None } else { *index };
 
     let index = if let Some(index) = index {
         scope.len() - index.get()
@@ -495,8 +488,6 @@ impl Engine {
         Self {
             packages: Default::default(),
             global_module: Default::default(),
-
-            #[cfg(not(feature = "no_module"))]
             module_resolver: None,
 
             type_names: HashMap::new(),
@@ -952,21 +943,24 @@ impl Engine {
         let mut idx_val = idx_values.pop();
 
         if is_index {
+            let pos = rhs.position();
+
             match rhs {
-                // xxx[idx].dot_rhs... | xxx[idx][dot_rhs]...
+                // xxx[idx].expr... | xxx[idx][expr]...
                 Expr::Dot(x) | Expr::Index(x) => {
+                    let (idx, expr, pos) = x.as_ref();
                     let is_idx = matches!(rhs, Expr::Index(_));
-                    let pos = x.0.position();
-                    let this_ptr = &mut self
-                        .get_indexed_mut(state, lib, obj, is_ref, idx_val, pos, op_pos, false)?;
+                    let idx_pos = idx.position();
+                    let this_ptr = &mut self.get_indexed_mut(
+                        state, lib, obj, is_ref, idx_val, idx_pos, op_pos, false,
+                    )?;
 
                     self.eval_dot_index_chain_helper(
-                        state, lib, this_ptr, &x.1, idx_values, is_idx, x.2, level, new_val,
+                        state, lib, this_ptr, expr, idx_values, is_idx, *pos, level, new_val,
                     )
                 }
                 // xxx[rhs] = new_val
                 _ if new_val.is_some() => {
-                    let pos = rhs.position();
                     let this_ptr = &mut self
                         .get_indexed_mut(state, lib, obj, is_ref, idx_val, pos, op_pos, true)?;
 
@@ -975,16 +969,7 @@ impl Engine {
                 }
                 // xxx[rhs]
                 _ => self
-                    .get_indexed_mut(
-                        state,
-                        lib,
-                        obj,
-                        is_ref,
-                        idx_val,
-                        rhs.position(),
-                        op_pos,
-                        false,
-                    )
+                    .get_indexed_mut(state, lib, obj, is_ref, idx_val, pos, op_pos, false)
                     .map(|v| (v.clone_into_dynamic(), false)),
             }
         } else {
@@ -1050,57 +1035,51 @@ impl Engine {
                     .map(|(v, _)| (v, false))
                 }
                 #[cfg(not(feature = "no_object"))]
-                // {xxx:map}.idx_lhs[idx_expr] | {xxx:map}.dot_lhs.rhs
+                // {xxx:map}.prop[expr] | {xxx:map}.prop.expr
                 Expr::Index(x) | Expr::Dot(x) if obj.is::<Map>() => {
+                    let (prop, expr, pos) = x.as_ref();
                     let is_idx = matches!(rhs, Expr::Index(_));
 
-                    let mut val = if let Expr::Property(p) = &x.0 {
+                    let mut val = if let Expr::Property(p) = prop {
                         let ((prop, _, _), _) = p.as_ref();
                         let index = prop.clone().into();
-                        self.get_indexed_mut(state, lib, obj, is_ref, index, x.2, op_pos, false)?
+                        self.get_indexed_mut(state, lib, obj, is_ref, index, *pos, op_pos, false)?
                     } else {
-                        // Syntax error
-                        return Err(Box::new(EvalAltResult::ErrorDotExpr(
-                            "".into(),
-                            rhs.position(),
-                        )));
+                        unreachable!();
                     };
 
                     self.eval_dot_index_chain_helper(
-                        state, lib, &mut val, &x.1, idx_values, is_idx, x.2, level, new_val,
+                        state, lib, &mut val, expr, idx_values, is_idx, *pos, level, new_val,
                     )
                 }
-                // xxx.idx_lhs[idx_expr] | xxx.dot_lhs.rhs
+                // xxx.prop[expr] | xxx.prop.expr
                 Expr::Index(x) | Expr::Dot(x) => {
+                    let (prop, expr, pos) = x.as_ref();
                     let is_idx = matches!(rhs, Expr::Index(_));
                     let args = &mut [obj, &mut Default::default()];
 
-                    let (mut val, updated) = if let Expr::Property(p) = &x.0 {
+                    let (mut val, updated) = if let Expr::Property(p) = prop {
                         let ((_, getter, _), _) = p.as_ref();
                         let args = &mut args[..1];
-                        self.exec_fn_call(state, lib, getter, true, 0, args, is_ref, None, x.2, 0)?
+                        self.exec_fn_call(state, lib, getter, true, 0, args, is_ref, None, *pos, 0)?
                     } else {
-                        // Syntax error
-                        return Err(Box::new(EvalAltResult::ErrorDotExpr(
-                            "".into(),
-                            rhs.position(),
-                        )));
+                        unreachable!();
                     };
                     let val = &mut val;
                     let target = &mut val.into();
 
                     let (result, may_be_changed) = self.eval_dot_index_chain_helper(
-                        state, lib, target, &x.1, idx_values, is_idx, x.2, level, new_val,
+                        state, lib, target, expr, idx_values, is_idx, *pos, level, new_val,
                     )?;
 
                     // Feed the value back via a setter just in case it has been updated
                     if updated || may_be_changed {
-                        if let Expr::Property(p) = &x.0 {
+                        if let Expr::Property(p) = prop {
                             let ((_, _, setter), _) = p.as_ref();
                             // Re-use args because the first &mut parameter will not be consumed
                             args[1] = val;
                             self.exec_fn_call(
-                                state, lib, setter, true, 0, args, is_ref, None, x.2, 0,
+                                state, lib, setter, true, 0, args, is_ref, None, *pos, 0,
                             )
                             .or_else(|err| match *err {
                                 // If there is no setter, no need to feed it back because the property is read-only
@@ -1127,13 +1106,16 @@ impl Engine {
         scope: &mut Scope,
         state: &mut State,
         lib: &FunctionsLib,
-        dot_lhs: &Expr,
-        dot_rhs: &Expr,
-        is_index: bool,
-        op_pos: Position,
+        expr: &Expr,
         level: usize,
         new_val: Option<Dynamic>,
     ) -> Result<Dynamic, Box<EvalAltResult>> {
+        let ((dot_lhs, dot_rhs, op_pos), is_index) = match expr {
+            Expr::Index(x) => (x.as_ref(), true),
+            Expr::Dot(x) => (x.as_ref(), false),
+            _ => unreachable!(),
+        };
+
         let idx_values = &mut StaticVec::new();
 
         self.eval_indexed_chain(scope, state, lib, dot_rhs, idx_values, 0, level)?;
@@ -1158,7 +1140,7 @@ impl Engine {
 
                 let this_ptr = &mut target.into();
                 self.eval_dot_index_chain_helper(
-                    state, lib, this_ptr, dot_rhs, idx_values, is_index, op_pos, level, new_val,
+                    state, lib, this_ptr, dot_rhs, idx_values, is_index, *op_pos, level, new_val,
                 )
                 .map(|(v, _)| v)
             }
@@ -1173,7 +1155,7 @@ impl Engine {
                 let val = self.eval_expr(scope, state, lib, expr, level)?;
                 let this_ptr = &mut val.into();
                 self.eval_dot_index_chain_helper(
-                    state, lib, this_ptr, dot_rhs, idx_values, is_index, op_pos, level, new_val,
+                    state, lib, this_ptr, dot_rhs, idx_values, is_index, *op_pos, level, new_val,
                 )
                 .map(|(v, _)| v)
             }
@@ -1489,14 +1471,14 @@ impl Engine {
                     Expr::Variable(_) => unreachable!(),
                     // idx_lhs[idx_expr] op= rhs
                     #[cfg(not(feature = "no_index"))]
-                    Expr::Index(x) => self.eval_dot_index_chain(
-                        scope, state, lib, &x.0, &x.1, true, x.2, level, new_val,
-                    ),
+                    Expr::Index(_) => {
+                        self.eval_dot_index_chain(scope, state, lib, lhs_expr, level, new_val)
+                    }
                     // dot_lhs.dot_rhs op= rhs
                     #[cfg(not(feature = "no_object"))]
-                    Expr::Dot(x) => self.eval_dot_index_chain(
-                        scope, state, lib, &x.0, &x.1, false, *op_pos, level, new_val,
-                    ),
+                    Expr::Dot(_) => {
+                        self.eval_dot_index_chain(scope, state, lib, lhs_expr, level, new_val)
+                    }
                     // Error assignment to constant
                     expr if expr.is_constant() => {
                         Err(Box::new(EvalAltResult::ErrorAssignmentToConstant(
@@ -1513,15 +1495,11 @@ impl Engine {
 
             // lhs[idx_expr]
             #[cfg(not(feature = "no_index"))]
-            Expr::Index(x) => {
-                self.eval_dot_index_chain(scope, state, lib, &x.0, &x.1, true, x.2, level, None)
-            }
+            Expr::Index(_) => self.eval_dot_index_chain(scope, state, lib, expr, level, None),
 
             // lhs.dot_rhs
             #[cfg(not(feature = "no_object"))]
-            Expr::Dot(x) => {
-                self.eval_dot_index_chain(scope, state, lib, &x.0, &x.1, false, x.2, level, None)
-            }
+            Expr::Dot(_) => self.eval_dot_index_chain(scope, state, lib, expr, level, None),
 
             #[cfg(not(feature = "no_index"))]
             Expr::Array(x) => Ok(Dynamic(Union::Array(Box::new(
@@ -1621,7 +1599,6 @@ impl Engine {
             }
 
             // Module-qualified function call
-            #[cfg(not(feature = "no_module"))]
             Expr::FnCall(x) if x.1.is_some() => {
                 let ((name, _, pos), modules, hash_fn_def, args_expr, def_val) = x.as_ref();
                 let modules = modules.as_ref().unwrap();
@@ -1642,7 +1619,7 @@ impl Engine {
                         .downcast_mut::<Module>()
                         .unwrap()
                 } else {
-                    scope.find_module(id).ok_or_else(|| {
+                    scope.find_module_internal(id).ok_or_else(|| {
                         Box::new(EvalAltResult::ErrorModuleNotFound(id.into(), *root_pos))
                     })?
                 };
@@ -1931,21 +1908,18 @@ impl Engine {
 
             // Import statement
             Stmt::Import(x) => {
-                #[cfg(feature = "no_module")]
-                unreachable!();
+                let (expr, (name, pos)) = x.as_ref();
 
-                #[cfg(not(feature = "no_module"))]
+                // Guard against too many modules
+                if state.modules >= self.max_modules {
+                    return Err(Box::new(EvalAltResult::ErrorTooManyModules(*pos)));
+                }
+
+                if let Some(path) = self
+                    .eval_expr(scope, state, lib, &expr, level)?
+                    .try_cast::<ImmutableString>()
                 {
-                    let (expr, (name, pos)) = x.as_ref();
-
-                    // Guard against too many modules
-                    if state.modules >= self.max_modules {
-                        return Err(Box::new(EvalAltResult::ErrorTooManyModules(*pos)));
-                    }
-
-                    if let Some(path) = self
-                        .eval_expr(scope, state, lib, &expr, level)?
-                        .try_cast::<ImmutableString>()
+                    #[cfg(not(feature = "no_module"))]
                     {
                         if let Some(resolver) = &self.module_resolver {
                             // Use an empty scope to create a module
@@ -1953,7 +1927,7 @@ impl Engine {
                                 resolver.resolve(self, Scope::new(), &path, expr.position())?;
 
                             let mod_name = unsafe_cast_var_name_to_lifetime(name, &state);
-                            scope.push_module(mod_name, module);
+                            scope.push_module_internal(mod_name, module);
 
                             state.modules += 1;
 
@@ -1964,9 +1938,12 @@ impl Engine {
                                 expr.position(),
                             )))
                         }
-                    } else {
-                        Err(Box::new(EvalAltResult::ErrorImportExpr(expr.position())))
                     }
+
+                    #[cfg(feature = "no_module")]
+                    Ok(Default::default())
+                } else {
+                    Err(Box::new(EvalAltResult::ErrorImportExpr(expr.position())))
                 }
             }