+++ src/module.rs
@@ -3,7 +3,7 @@
 use crate::any::{Dynamic, Variant};
 use crate::calc_fn_hash;
 use crate::engine::{make_getter, make_setter, Engine, FunctionsLib, FUNC_INDEXER};
-use crate::fn_native::{CallableFunction, FnCallArgs, IteratorFn};
+use crate::fn_native::{CallableFunction, FnCallArgs, IteratorFn, SendSync};
 use crate::parser::{
     FnAccess,
     FnAccess::{Private, Public},
@@ -12,7 +12,7 @@ use crate::parser::{
 use crate::result::EvalAltResult;
 use crate::scope::{Entry as ScopeEntry, EntryType as ScopeEntryType, Scope};
 use crate::token::{Position, Token};
-use crate::utils::StaticVec;
+use crate::utils::{StaticVec, StraightHasherBuilder};
 
 use crate::stdlib::{
     any::TypeId,
@@ -44,10 +44,14 @@ pub struct Module {
     variables: HashMap<String, Dynamic>,
 
     /// Flattened collection of all module variables, including those in sub-modules.
-    all_variables: HashMap<u64, Dynamic>,
+    all_variables: HashMap<u64, Dynamic, StraightHasherBuilder>,
 
     /// External Rust functions.
-    functions: HashMap<u64, (String, FnAccess, StaticVec<TypeId>, CallableFunction)>,
+    functions: HashMap<
+        u64,
+        (String, FnAccess, StaticVec<TypeId>, CallableFunction),
+        StraightHasherBuilder,
+    >,
 
     /// Script-defined functions.
     lib: FunctionsLib,
@@ -57,7 +61,7 @@ pub struct Module {
 
     /// Flattened collection of all external Rust functions, native or scripted,
     /// including those in sub-modules.
-    all_functions: HashMap<u64, CallableFunction>,
+    all_functions: HashMap<u64, CallableFunction, StraightHasherBuilder>,
 }
 
 impl fmt::Debug for Module {
@@ -101,7 +105,7 @@ impl Module {
     /// ```
     pub fn new_with_capacity(capacity: usize) -> Self {
         Self {
-            functions: HashMap::with_capacity(capacity),
+            functions: HashMap::with_capacity_and_hasher(capacity, StraightHasherBuilder),
             ..Default::default()
         }
     }
@@ -941,23 +945,7 @@ impl ModuleRef {
 }
 
 /// Trait that encapsulates a module resolution service.
-#[cfg(not(feature = "no_module"))]
-#[cfg(not(feature = "sync"))]
-pub trait ModuleResolver {
-    /// Resolve a module based on a path string.
-    fn resolve(
-        &self,
-        engine: &Engine,
-        scope: Scope,
-        path: &str,
-        pos: Position,
-    ) -> Result<Module, Box<EvalAltResult>>;
-}
-
-/// Trait that encapsulates a module resolution service.
-#[cfg(not(feature = "no_module"))]
-#[cfg(feature = "sync")]
-pub trait ModuleResolver: Send + Sync {
+pub trait ModuleResolver: SendSync {
     /// Resolve a module based on a path string.
     fn resolve(
         &self,
@@ -975,6 +963,8 @@ pub mod resolvers {
     pub use super::file::FileModuleResolver;
     pub use super::stat::StaticModuleResolver;
 }
+#[cfg(feature = "no_module")]
+pub mod resolvers {}
 
 /// Script file-based module resolver.
 #[cfg(not(feature = "no_module"))]